13.13. *Implementar as consultas com dados agregados (solução pessoal)
13.13. Discutindo sobre onde implementar as consultas com dados
agregados

+++ gorisfood-api/src/main/java/br/com/felipe/gorisfood/domain/model/projection/VendaDiaria.java
@@ -13,6 +13,6 @@ import lombok.Setter;
 public class VendaDiaria {
 	
 	private LocalDate data;
-	private Integer totalVendas;
+	private Long totalVendas;
 	private BigDecimal totalFaturado;
 }
